Implemented the no commits test for Android.
Required some moving around, particularly removing the background step in the feature file, so we would have the chance to force the data we want in the mock backend before launching the app.

Minor readability issue that maaaaaassively simplifies the technical implementation. The impact on readibility is actually even debatable!