Now hardcoding scheme and build dir inside the XCode build command.
This is the *opposite* of what we thought we would as per #11 and #13. The truth is that this is not a tutorial about how to compile mobile apps from the terminal.

This it not a project template that you can clone, change a few configuration values and get the project up and running quickly.

This project is a reference or guide about our BDD set up.

Nothing more, nothing less!